Fix win condition clauses
The win check clauses were fixed so that it looks at the active player's token for a possible win. If the board is full, the game ends in a draw. In addition, the user is forced to choose a place for the token that has not been used to break the loop.